Move run() out of lib.rs into dedicated module
AGENTS.md specifies that lib.rs should only contain API docs and

re-exports — no logic or function definitions. Move the run() entry

point into a new src/run.rs module and re-export it from lib.rs.
